Adopt existing test style.
